Refactor WooPosSettingsScreen to use WooPosSettingsViewModel and add new ViewModel for settings navigation state management
