+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/settings/WooPosSettingsViewModel.kt
@@ -1,9 +1,6 @@
-package com.woocommerce.android.ui.woopos.settings.container
+package com.woocommerce.android.ui.woopos.settings
 
 import androidx.lifecycle.ViewModel
-import com.woocommerce.android.ui.woopos.settings.WooPosSettingsState
-import com.woocommerce.android.ui.woopos.settings.SettingsCategory
-import com.woocommerce.android.ui.woopos.settings.SettingsDetailDestination
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -12,11 +9,11 @@ import kotlinx.coroutines.flow.update
 import javax.inject.Inject
 
 @HiltViewModel
-class WooPosSettingsContainerViewModel @Inject constructor() : ViewModel() {
-    
+class WooPosSettingsViewModel @Inject constructor() : ViewModel() {
+
     private val _navigationState = MutableStateFlow(WooPosSettingsState())
     val navigationState: StateFlow<WooPosSettingsState> = _navigationState.asStateFlow()
-    
+
     fun onCategorySelected(category: SettingsCategory) {
         _navigationState.update { currentState ->
             val newDetailDestination = when (category) {
@@ -28,7 +25,7 @@ class WooPosSettingsContainerViewModel @Inject constructor() : ViewModel() {
             )
         }
     }
-    
+
     fun navigateToDetail(destination: SettingsDetailDestination) {
         _navigationState.update { currentState ->
             currentState.copy(
@@ -36,7 +33,7 @@ class WooPosSettingsContainerViewModel @Inject constructor() : ViewModel() {
             )
         }
     }
-    
+
     fun popDetailBackStack() {
         _navigationState.update { currentState ->
             if (currentState.detailBackStack.size > 1) {
